+++ CHANGELOG.md
@@ -10,6 +10,7 @@ HumHub Changelog
 - Fix #4940: Typo in German translation "Benachrichtigungen"
 - Fix #4942: MemberOf Display in LDAP CLI Show User Details
 - Fix #4465: LDAP PHP 8 incompatibility (multiPageSearch)
+- Enh #4569: Config to overwrite module path
 
 
 1.8.0 (March 1, 2021)